refactor: restructure distributors module by separating controller, service, and model files
