+++ apps/api/src/controllers/v1/distributors/distributors.model.ts
@@ -1,0 +1,3 @@
+import { distributorsTable } from "../../../db/schemas/distributors.drizzle";
+
+export { distributorsTable };
